requirements.txt
tensorflow>=2.12.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
pandas>=1.3.0
nibabel>=3.2.0
scipy>=1.7.0
tqdm>=4.62.0
argparse